stricter type checking on oh_item, fix datetimeaddDateComparisonSupport
