+++ bundles/org.openhab.ui/web/src/assets/definitions/blockly/utils.js
@@ -116,9 +116,10 @@ export function addChrono () {
   return chrono
 }
 
-export function addDateComparisonSupport () {
+export function addDateComparisonSupportNashorn () {
   let zdtCompare = javascriptGenerator.provideFunction_(
     'zdtCompare', [
+      '// Nashorn',
       'function ' + javascriptGenerator.FUNCTION_NAME_PLACEHOLDER_ + '(zdt1, zdt2, compareOp, precision, compDate) {',
       '  switch (precision) {',
       '    case \'years\':',
@@ -157,3 +158,47 @@ export function addDateComparisonSupport () {
     ])
   return zdtCompare
 }
+
+export function addDateComparisonSupportGraalVM () {
+  const graalZdtCompare = javascriptGenerator.provideFunction_(
+      'zdtCompare', [
+        '// graalVM',
+        'function ' + javascriptGenerator.FUNCTION_NAME_PLACEHOLDER_ + '(zdt1, zdt2, compareOp, precision, compDate) {',
+        '  switch (precision) {',
+        '    case \'years\':',
+        '     zdt2 = zdt2.withMonth(zdt1.monthValue());',
+        '    case \'months\':',
+        '     zdt2 = zdt2.withDayOfMonth(zdt1.dayOfMonth());',
+        '    case \'days\':',
+        '     zdt2 = zdt2.withHour(zdt1.hour());',
+        '    case \'hours\':',
+        '     zdt2 = zdt2.withMinute(zdt1.minute());',
+        '    case \'minutes\':',
+        '     zdt2 = zdt2.withSecond(zdt1.second());',
+        '    case \'seconds\':',
+        '     zdt2 = zdt2.withNano(zdt1.nano());',
+        '  }',
+        '  if (compDate === \'date\') {',
+        '    zdt1 = zdt1.toLocalDate();',
+        '    zdt2 = zdt2.toLocalDate();',
+        '  } else if (compDate === \'time\') {',
+        '    zdt1 = zdt1.toLocalTime();',
+        '    zdt2 = zdt2.toLocalTime();',
+        '  }',
+        '  switch (compareOp) {',
+        '    case \'before\':',
+        '      return zdt1.isBefore(zdt2);',
+        '    case \'equal\':',
+        '      return zdt1.equals(zdt2);',
+        '    case \'after\':',
+        '      return zdt1.isAfter(zdt2);',
+        '    case \'beforeEqual\':',
+        '      return zdt1.isBefore(zdt2) || zdt1.equals(zdt2);',
+        '    case \'afterEqual\':',
+        '      return zdt1.isAfter(zdt2) || zdt1.equals(zdt2);',
+        '  }',
+        '}'
+      ])
+  return graalZdtCompare
+}
+
