+++ bundles/org.openhab.ui/web/src/pages/settings/rules/script/blockly-editor.vue
@@ -384,37 +384,37 @@
             </value>
           </block>
           <block type="oh_getitem_attribute">
-            <value name="item">
-              <shadow type="oh_getitem">
-                <value name="itemName">
-                  <shadow type="oh_item" />
-                </value>
-              </shadow>
-            </value>
-          </block>
-          <block type="oh_getitem">
-            <value name="itemName">
-              <shadow type="oh_item" />
-            </value>
-          </block>
-          <block type="oh_groupmembers">
-            <value name="groupName">
-              <shadow type="oh_item" />
-            </value>
-          </block>
-          <block type="oh_taggeditems">
-            <value name="tagName">
-              <shadow type="text" />
-            </value>
-          </block>
-          <block type="oh_event">
-            <value name="value">
-              <shadow type="text">
-                <field name="TEXT">value</field>
-              </shadow>
-            </value>
-            <value name="itemName">
-              <shadow type="oh_item" />
+             <value name="item">
+               <shadow type="oh_getitem">
+                 <value name="itemName">
+                   <shadow type="oh_item" />
+                 </value>
+               </shadow>
+             </value>
+           </block>
+           <block type="oh_getitem">
+             <value name="itemName">
+               <shadow type="oh_item" />
+             </value>
+           </block>
+           <block type="oh_groupmembers">
+             <value name="groupName">
+               <shadow type="oh_item" />
+             </value>
+           </block>
+           <block type="oh_taggeditems">
+             <value name="tagName">
+               <shadow type="text" />
+             </value>
+           </block>
+           <block type="oh_event">
+             <value name="value">
+               <shadow type="text">
+                 <field name="TEXT">value</field>
+               </shadow>
+             </value>
+             <value name="itemName">
+               <shadow type="oh_item" />
             </value>
           </block>
           <block type="oh_item" />
@@ -616,69 +616,69 @@
             </value>
           </block>
           <sep gap="48" />
-          <block type="oh_zdt_amend">
-            <value name="baseZdt">
-              <shadow type="oh_zdt" />
-            </value>
-          </block>
-          <block type="oh_zdt_temporal_unit" />
-          <block type="oh_zdt_temporal_unit_input" />
-          <sep gap="48" />
+         <block type="oh_zdt_amend">
+           <value name="baseZdt">
+             <shadow type="oh_zdt" />
+           </value>
+         </block>
+         <block type="oh_zdt_temporal_unit" />
+         <block type="oh_zdt_temporal_unit_input" />
+         <sep gap="48" />
           <block type="oh_zdt_fromItem">
-            <value name="itemName">
-              <shadow type="oh_item" />
-            </value>
+              <value name="itemName">
+                <shadow type="oh_item" />
+              </value>
           </block>
           <block type="oh_zdt_toText">
             <value name="date">
               <shadow type="oh_zdt" />
             </value>
           </block>
-          <sep gap="48" />
-          <block type="oh_zdt_compare">
-            <value name="zdtOne">
-              <shadow type="oh_zdt" />
-            </value>
-            <value name="zdtTwo">
-              <shadow type="oh_zdt" />
-            </value>
-          </block>
-          <block type="oh_zdt_between">
-            <value name="zdtOne">
-              <shadow type="oh_zdt" />
-            </value>
-            <value name="zdtTwo">
-              <shadow type="oh_zdt" />
-            </value>
-            <value name="zdtThree">
-              <shadow type="oh_zdt" />
-            </value>
-          </block>
-          <sep gap="48" />
-          <block type="oh_get_zdt_part">
-            <value name="date">
-              <shadow type="oh_zdt" />
-            </value>
-          </block>
-          <block type="oh_get_time_between">
-            <value name="zdtOne">
-              <shadow type="oh_zdt" />
-            </value>
-            <value name="zdtTwo">
-              <shadow type="oh_zdt" />
-            </value>
-          </block>
-          <sep gap="48" />
-          <block type="oh_dayoffset_today" />
-          <block type="oh_dayoffset">
-            <value name="offset">
-              <shadow type="math_number">
-                <field name="NUM">
-                  0
-                </field>
-              </shadow>
-            </value>
-          </block>
+         <sep gap="48" />
+         <block type="oh_zdt_compare">
+           <value name="zdtOne">
+             <shadow type="oh_zdt" />
+           </value>
+           <value name="zdtTwo">
+             <shadow type="oh_zdt" />
+           </value>
+         </block>
+         <block type="oh_zdt_between">
+           <value name="zdtOne">
+             <shadow type="oh_zdt" />
+           </value>
+           <value name="zdtTwo">
+             <shadow type="oh_zdt" />
+           </value>
+           <value name="zdtThree">
+             <shadow type="oh_zdt" />
+           </value>
+         </block>
+         <sep gap="48" />
+         <block type="oh_get_zdt_part">
+           <value name="date">
+             <shadow type="oh_zdt" />
+           </value>
+         </block>
+         <block type="oh_get_time_between">
+           <value name="zdtOne">
+             <shadow type="oh_zdt" />
+           </value>
+           <value name="zdtTwo">
+             <shadow type="oh_zdt" />
+           </value>
+         </block>
+         <sep gap="48" />
+         <block type="oh_dayoffset_today" />
+         <block type="oh_dayoffset">
+           <value name="offset">
+             <shadow type="math_number">
+               <field name="NUM">
+                 0
+               </field>
+             </shadow>
+           </value>
+         </block>
         </category>
 
         <category name="Ephemeris">
